+++ examples/basic/src/model-provider/elevenlabs/elevenlabs-stream-speech-example.ts
@@ -22,6 +22,7 @@ async function main() {
   const speechStream = await streamSpeech(
     new ElevenLabsSpeechModel({
       voice: "pNInz6obpgDQGcFmaJgB", // Adam
+      optimizeStreamingLatency: 1,
       voiceSettings: {
         stability: 1,
         similarityBoost: 0.35,